Test implementation of party mode
If party mode is on, the app will track notifications for spotify links and start playing the linked song automatically. Lots of bugs currently.

A big one is that if someone links a song, and someone else sends you a message before the person who linked the song sends anything, the linked song starts playing again because the notification isn't deleted. -----> implement some kind of ability to delete the notifications OR check if the song is already in the queue when adding it